Rename ProviderBasedPrivateKeyReader interface.
For compatibility reasons, I originally created the interface for
private key reading operations and named it
ProviderBasedPrivateKeyReader and then had PrivateKeyReader implement
it.  However, these names are backwards from what one would expect.
This patch corrects the confusion in naming.
